+++ CHANGELOG.md
@@ -637,7 +637,7 @@ st.write("Hello,", name or "world")
 
 #### Changed
 
-- New Stite and Stlite Sharing logos, [#1130](https://github.com/whitphx/stlite/pull/1130).
+- New Stlite and Stlite Sharing logos, [#1130](https://github.com/whitphx/stlite/pull/1130).
 
 ## [0.65.1] - 2024-09-17